Reorganize official_products into generic/kepler/tess modules
